data/utilities/snowflake-cloner/pyproject.toml
[tool.poetry]
name = "snowflakecloner-ext"
version = "0.0.1"
description = "`snowflakecloner-ext` is a Meltano utility extension."
authors = ["Pat Nadolny"]
license = "Apache 2.0"
packages = [
    { include = "snowflakecloner_ext" }
    # uncomment this if you want to include static assets
    # { include = "files_snowflakecloner-ext.*" }
]
# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_snowflakecloner-ext/"
#include = [
#    "files_snowflakecloner-ext/somefile.thing"
#]

[tool.poetry.dependencies]
python = "<3.12,>=3.7"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk"= {git = "https://github.com/meltano/edk.git", rev="main"}
tqdm = "^4.65.0"
snowflake-connector-python = "^3.0.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
snowflakecloner_extension = 'snowflakecloner_ext.main:app'
snowflakecloner_invoker = 'snowflakecloner_ext.pass_through:pass_through_cli'
